requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
requests==2.28.2
python-dotenv==1.0.0
Werkzeug==2.0.3